+++ src/axom/primal/tests/primal_numeric_array.cpp
@@ -7,8 +7,8 @@
 
 #include "axom/primal/geometry/NumericArray.hpp"
 
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+#include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 using namespace axom;
 
@@ -254,7 +254,7 @@ int main(int argc, char* argv[])
 
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
 
   // finalized when exiting main scope