+++ src/axom/quest/tests/quest_signed_distance.cpp
@@ -5,8 +5,8 @@
 
 #include "axom/config.hpp"
 #include "axom/slic/interface/slic.hpp"
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+#include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 #include "axom/mint/mesh/CellTypes.hpp"
 #include "axom/mint/config.hpp"
@@ -194,7 +194,7 @@ int main(int argc, char* argv[])
 
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
 
   // finalized when exiting main scope