+++ src/axom/slam/examples/tinyHydro/tests/slam_tinyHydro_sod1DTwoPart.cpp
@@ -156,14 +156,14 @@ void tinyHydroSod1D_2part()
 }
 //----------------------------------------------------------------------
 //----------------------------------------------------------------------
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+#include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 int main()
 {
   int result = 0;
 
-  UnitTestLogger logger;  // create & initialize test logger,
+  SimpleLogger logger;  // create & initialize test logger,
 
   // finalized when exiting main scope
   tinyHydroSod1D_2part();