+++ src/axom/slic/tests/slic_benchmark_asserts.cpp
@@ -5,8 +5,8 @@
 
 #include "benchmark/benchmark_api.h"
 #include "axom/slic/interface/slic.hpp"
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
+#include "axom/slic/core/SimpleLogger.hpp"
+using axom::slic::SimpleLogger;
 
 /*!
  * \file
@@ -187,7 +187,7 @@ int main(int argc, char* argv[])
   printMsg("at start of main");
 
   printMsg(" Before init logger");
-  axom::slic::UnitTestLogger logger;  // create & initialize test logger,
+  axom::slic::SimpleLogger logger;  // create & initialize test logger,
   printMsg(" After init logger");
 
   printMsg(" Before init benchmark");