+++ src/axom/slic/tests/slic_fmt.cpp
@@ -13,9 +13,9 @@
 #include "fmt/fmt.hpp"
 
 #include "axom/slic/interface/slic.hpp"
-#include "axom/slic/core/UnitTestLogger.hpp"
+#include "axom/slic/core/SimpleLogger.hpp"
 
-using axom::slic::UnitTestLogger;
+using axom::slic::SimpleLogger;
 
 #include "gtest/gtest.h"
 
@@ -53,8 +53,8 @@ int main(int argc, char* argv[])
 
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
-                          // finalized when exiting main scope
+  SimpleLogger logger;  // create & initialize test logger,
+                        // finalized when exiting main scope
 
   result = RUN_ALL_TESTS();